DIS-37 Phase 4 review fixes: re-export type, move workspace-prep, rename
Address all four items from the round-1 review:

#1268 (high) — `WorktreeStatus` was duplicated, not re-exported. Replaced
the parallel definition in `agents/types.ts:62-70` with
`export type { WorktreeStatus } from "../shared/git/worktree-status.js"`.
The shared module is now the single source of truth; manager's public
re-export keeps working transparently.

#1269 (medium) — `setupWorktree` was misplaced in `shared/git/worktree.ts`
(it's not a git operation — it copies `.env` and shells out to a package
manager). Moved to a new module `agents/workspace-prep.ts` and renamed
to `setupAgentWorkspace` to make the agent-flow-specific intent obvious
(its sibling, the tmux-mode launch path's bash equivalent, lives in
`agents/tmux/setup-script.ts`). Removed the now-orphaned `copyFile`,
`stat`, and `FastifyBaseLogger` imports from `worktree.ts` — that file
is back to being pure git porcelain.

#1270 (low) — Dropped the orphaned `copyFile` import from
`agents/manager.ts:3`. Audited the remaining `fs/promises` imports
(`mkdir`, `readFile`, `rm`, `stat`, `unlink`, `writeFile`) — all 6 still
have live call sites in manager.ts.

#1271 (low) — Renamed the shared inspector
`checkWorktreeStatus(worktreePath)` to
`readWorktreeStatus(worktreePath)`. The path-keyed name now distinguishes
it from the manager method's id-keyed `checkWorktreeStatus(id)` façade
without an import alias. The aliased `as inspectWorktreeStatus` in
manager.ts is gone.

Test plumbing:
- `test/worktree-setup.test.ts` → `test/workspace-prep.test.ts`
  (renamed via git mv to keep history)
- All `setupWorktree` references in the test → `setupAgentWorkspace`
- Import path updated to `../src/agents/workspace-prep.js`
- `test/worktree-status.test.ts` updated for the rename

754/762 server vitest pass (unchanged). Typecheck green.
